backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
praw>=7.8.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
openai>=1.54.0
